// examples/multipackage/client/src/rpc/types.ts
// Code generated by tygor. DO NOT EDIT.

export interface v1_GetUserRequest {
  id: number;
}
export interface v2_GetUserRequest {
  id: number;
}
export interface MigrationRequest {
  v1_user: v1_User;
  v2_user: v2_User;
}
export interface MigrationResponse {
  success: boolean;
  v1_user: v1_User;
  v2_user: v2_User;
}
export interface v1_User {
  id: number;
  name: string;
}
export interface v2_User {
  id: number;
  name: string;
  email: string;
  created_at: string;
}